pyproject.toml
[project]
name = "cvm-demo"
version = "0.1.0"
description = "Local-first AI-native harness MVP"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "cvm-platform-api",
  "cvm-temporal-worker",
  "cvm-eval-runner",
]

[dependency-groups]
dev = [
  "datamodel-code-generator>=0.31,<0.32",
  "httpx>=0.28,<0.29",
  "pytest>=8.3,<9",
  "pytest-cov>=5,<6",
  "pyyaml>=6,<7",
  "ruff>=0.11,<0.12",
]

[tool.uv.sources]
cvm-platform-api = { workspace = true }
cvm-temporal-worker = { workspace = true }
cvm-eval-runner = { workspace = true }
cvm-contracts-generated = { workspace = true }
cvm-domain-kernel = { workspace = true }
cvm-testkit = { workspace = true }

[tool.uv.workspace]
members = [
  "services/platform-api",
  "services/temporal-worker",
  "services/eval-runner",
  "libs/py/contracts-generated",
  "libs/py/domain-kernel",
  "libs/py/testkit",
]

[tool.pytest.ini_options]
pythonpath = [
  "services/platform-api/src",
  "services/temporal-worker/src",
  "services/eval-runner/src",
  "libs/py/contracts-generated/src",
  "libs/py/domain-kernel/src",
  "libs/py/testkit/src",
]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
